+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
 
 - Recipe "Translating content from URL"
+- Predefined Persisted Query "Translate content from URL"
 
 ### Fixed