+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Recipe "Translating content from URL"
 - Predefined Persisted Query "Translate content from URL"
 - `XML` scalar type
+- Added documentation for new field `_strDecodeXMLAsJSON` from the Helper Function Collection extension
 
 ### Fixed