+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -9,7 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
 
 - `XML` scalar type
-- Added documentation for new field `_strDecodeXMLAsJSON` from the Helper Function Collection extension
+- Documentation for new field `_strDecodeXMLAsJSON` from the Helper Function Collection extension
 - Recipe "Translating content from URL"
 - Predefined Persisted Query "Translate content from URL"
 - Predefined Persisted Query "Import post from WordPress RSS feed"