+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Documentation for new field `_strDecodeXMLAsJSON` from the Helper Function Collection extension
 - Documentation for new field `_strParseCSV` from the Helper Function Collection extension
 - Recipe "Translating content from URL"
+- Persisted Queries "Translate post (Classic editor)" and "Translate posts (Classic editor)"
 - Predefined Persisted Query "Translate content from URL"
 - Predefined Persisted Query "Import post from WordPress RSS feed"
 - Predefined Persisted Query "Import posts from CSV"