+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 1.2.0 - DATE
+## 1.3.0 - DATE
+
+## 1.2.0 - 28/11/2013
 
 ### Added